requirements.txt
flask
ruamel.yaml
mysql-connector-python-rf